requirements.txt
streamlit==1.42.2
openpyxl==3.1.5
pandas==2.2.3
pywebview==5.4
